+++ src/Oro/Bundle/MessageQueueBundle/Tests/Functional/StatusCalculator/QueryCalculatorTest.php
@@ -30,7 +30,7 @@ class QueryCalculatorTest extends WebTestCase
     public function testCalculateRootJobProgressForAsyncChild(): void
     {
         /** @var Job $job */
-        $job = $this->getReference(LoadJobData::JOB_5);
+        $job = $this->getJobReference(LoadJobData::JOB_5);
         $this->assertInstanceOf(PersistentCollection::class, $job->getChildJobs());
         $this->assertFalse($job->getChildJobs()->isInitialized());
 
@@ -57,7 +57,7 @@ class QueryCalculatorTest extends WebTestCase
     public function testCalculateRootJobStatusForAsyncChild(): void
     {
         /** @var Job $job */
-        $job = $this->getReference(LoadJobData::JOB_5);
+        $job = $this->getJobReference(LoadJobData::JOB_5);
         $this->assertInstanceOf(PersistentCollection::class, $job->getChildJobs());
         $this->assertFalse($job->getChildJobs()->isInitialized());
 
@@ -119,5 +119,13 @@ class QueryCalculatorTest extends WebTestCase
         $child->setRootJob($job);
         $child->setCreatedAt(new \DateTime());
         $jobManager->saveJob($child);
+    }
+
+    protected function getJobReference(string $name): ?Job
+    {
+        return $this->getContainer()
+            ->get('oro_entity.doctrine_helper')
+            ->getEntityRepository(Job::class)
+            ->findOneBy(['name' => $name]);
     }
 }