+++ src/Oro/Bundle/UserBundle/Tests/Functional/Acl/AbstractPermissionConfigurableTestCase.php
@@ -20,7 +20,7 @@ abstract class AbstractPermissionConfigurableTestCase extends WebTestCase
     }
 
     /**
-     * @afterClass
+     * @beforeResetClient
      */
     public static function buildOriginCache()
     {